requirements.txt
langgraph
langchain_community
langchain
python-dotenv
ipykernel
langchain-openai